refactor: output 로직 scoreboard 객체에 위임
